% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabexp.R
\name{tabexp}
\alias{tabexp}
\title{Compute events and exposure  by years before survey (YBS)
from DHS-like data using person-period file logic.}
\usage{
tabexp(
  data = NULL,
  ybs_max = 40,
  age_bins = seq(10, 50, by = 5),
  birth_vars = NULL,
  default_vars = c("caseid", "v011", "v008", "v005", "awfactt")
)
}
\arguments{
\item{data}{A data.frame or data.table with DHS variables.}

\item{ybs_max}{Maximum years before survey to include.}

\item{age_bins}{Age group breaks.}

\item{birth_vars}{Names of birth date variables (e.g., b3_01 to b3_20).}

\item{default_vars}{Names of essential variables (e.g., v011, v008, v005).}
}
\value{
A data.table with YBS, age group, events, exposure, and mid-period year.
}
\description{
Compute events and exposure  by years before survey (YBS)
from DHS-like data using person-period file logic.
}
